Refactor: rename server actions
